+++ docs/changelog.md
@@ -34,6 +34,7 @@ Ends August 17, 2020
 - add buildspec for AWS CodeBuild.
 - update buildspec so that testing closes onces finished.
 - prevent Hand component from hiding when it's clicked.
+- hide Hand component if it's already displaying and a Card compnent is clicked.
 
 ## Sprint 1
 Ended August 3, 2020